small refactors and add new event listener for ios
